feature: support initrd secrets
